chore: move navigation events from auto height to ad
